nofity api, use float64 to store balance.

+++ pkg/remotetypes.go
@@ -55,7 +55,7 @@ type apiBalance struct {
 	Namespace string  `json:"namespace"`
 	CreateAt  string  `json:"create_at"`
 	UpdateAt  string  `json:"update_at"`
-	Balance   float32 `json:"balance"`
+	Balance   float64 `json:"balance"`
 	Status    string  `json:"state,omitempty"`
 }
